Refactored filenames and added router
