+++ app/src/components/Dashboard.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Dashboard = () => {
+    return (
+        <div>You're at the dashboard</div>
+    )
+}
+
+export default Dashboard
